talks/01_static_islands_dynamic_sea/requirements.txt
django>=5.1
django-mantle
django-mantle-drf
djangorestframework
drf-spectacular
jsonschema-extractor
